Update README.md
This is a repository about the steps folowed in data mining with a focus on dataset provided for us to analyse as agroup. The dataset was in a csv form and we had to extract and load it before we start using the various tools in python to do the analysis.